Refactor trade-producer service to introduc an abstract base class for Kraken API to support pluggable APIs.
